Move tests into top-level directory
